Reduced memory footprint of HPLC-DAD chromatograms

+++ chemclipse/plugins/org.eclipse.chemclipse.ux.extension.xxd.ui/src/org/eclipse/chemclipse/ux/extension/xxd/ui/internal/provider/ScanTableComparator.java
@@ -179,7 +179,7 @@ public class ScanTableComparator extends AbstractRecordTableComparator implement
 			//
 			switch(getPropertyIndex()) {
 				case 0:
-					sortOrder = Double.compare(scanSignalWSD2.getWavelength(), scanSignalWSD1.getWavelength());
+					sortOrder = Float.compare(scanSignalWSD2.getWavelength(), scanSignalWSD1.getWavelength());
 					break;
 				case 1:
 					sortOrder = Float.compare(scanSignalWSD2.getAbundance(), scanSignalWSD1.getAbundance());
