+++ chemclipse/plugins/org.eclipse.chemclipse.wsd.model/src/org/eclipse/chemclipse/wsd/model/core/interpolation/RasterizeCalculator.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2022 Lablicate GmbH.
+ * Copyright (c) 2022, 2023 Lablicate GmbH.
  *
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v1.0
@@ -25,7 +25,7 @@ public class RasterizeCalculator {
 	public static final int MIN_STEPS = 1;
 	public static final int MAX_STEPS = 10;
 
-	public static Map<Integer, Float> apply(TreeMap<Double, Float> dataOriginal, int steps) {
+	public static Map<Integer, Float> apply(TreeMap<Float, Float> dataOriginal, int steps) {
 
 		Map<Integer, Float> dataAdjusted = null;
 		if(steps >= RasterizeCalculator.MIN_STEPS && steps <= RasterizeCalculator.MAX_STEPS) {
@@ -34,19 +34,19 @@ public class RasterizeCalculator {
 				 * Range
 				 */
 				dataAdjusted = new HashMap<>();
-				Entry<Double, Float> firstWavelengthEntry = dataOriginal.firstEntry();
-				Entry<Double, Float> lastWavelengthEntry = dataOriginal.lastEntry();
+				Entry<Float, Float> firstWavelengthEntry = dataOriginal.firstEntry();
+				Entry<Float, Float> lastWavelengthEntry = dataOriginal.lastEntry();
 				//
-				int startWavelengh = (int)Math.round(firstWavelengthEntry.getKey());
-				adjust(dataAdjusted, startWavelengh, firstWavelengthEntry, dataOriginal.ceilingEntry((double)(startWavelengh + 1)));
+				int startWavelengh = Math.round(firstWavelengthEntry.getKey());
+				adjust(dataAdjusted, startWavelengh, firstWavelengthEntry, dataOriginal.ceilingEntry((float)(startWavelengh + 1)));
 				//
-				int stopWavelengh = (int)Math.round(lastWavelengthEntry.getKey());
-				adjust(dataAdjusted, stopWavelengh, dataOriginal.floorEntry((double)(stopWavelengh - 1)), lastWavelengthEntry);
+				int stopWavelengh = Math.round(lastWavelengthEntry.getKey());
+				adjust(dataAdjusted, stopWavelengh, dataOriginal.floorEntry((float)(stopWavelengh - 1)), lastWavelengthEntry);
 				//
 				for(int wavelength = (startWavelengh + 1); wavelength <= (stopWavelengh - 1); wavelength++) {
 					if(wavelength % steps == 0) {
-						Entry<Double, Float> floorEntry = dataOriginal.floorEntry((double)wavelength);
-						Entry<Double, Float> ceilingEntry = dataOriginal.ceilingEntry((double)wavelength);
+						Entry<Float, Float> floorEntry = dataOriginal.floorEntry((float)wavelength);
+						Entry<Float, Float> ceilingEntry = dataOriginal.ceilingEntry((float)wavelength);
 						adjust(dataAdjusted, wavelength, floorEntry, ceilingEntry);
 					}
 				}
@@ -56,7 +56,7 @@ public class RasterizeCalculator {
 		return dataAdjusted;
 	}
 
-	private static void adjust(Map<Integer, Float> dataAdjusted, int wavelength, Entry<Double, Float> floorEntry, Entry<Double, Float> ceilingEntry) {
+	private static void adjust(Map<Integer, Float> dataAdjusted, int wavelength, Entry<Float, Float> floorEntry, Entry<Float, Float> ceilingEntry) {
 
 		Point p1 = new Point(floorEntry.getKey(), floorEntry.getValue());
 		Point p2 = new Point(ceilingEntry.getKey(), ceilingEntry.getValue());
