+++ chemclipse/plugins/org.eclipse.chemclipse.wsd.model/src/org/eclipse/chemclipse/wsd/model/xwc/IExtractedSingleWavelengthSignals.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2017, 2018 Lablicate GmbH.
+ * Copyright (c) 2017, 2023 Lablicate GmbH.
  *
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v1.0
@@ -32,5 +32,5 @@ public interface IExtractedSingleWavelengthSignals extends ITotalScanSignals {
 	@Override
 	IExtractedSingleWavelengthSignals makeDeepCopy();
 
-	double getWavelength();
+	float getWavelength();
 }
